Refactor lead conversion reporting
